# fix: do not modify iter during List.map and Map.map (#1649)