--- a/src/Map.js
+++ b/src/Map.js
@@ -128,7 +128,7 @@ export class Map extends KeyedCollection {
   map(mapper, context) {
     return this.withMutations(map => {
       map.forEach((value, key) => {
-        map.set(key, mapper.call(context, value, key, map));
+        map.set(key, mapper.call(context, value, key, this));
       });
     });
   }