@Methuselah96 2020-08-15
@mishawakerman This looks great! Can you resubmit this PR on https://github.com/immutable-js-oss/immutable-js? We're moving maintenance to that repo until we can get access here.