@mishawakerman 2020-08-16
Sure thing. I've seen some alerts regarding lee byron (finally!) handing the project over, is there an update about the move to immutable-js-oss I can read somewhere?