@Methuselah96 2020-08-17
@mishawakerman No, @leebyron hasn't handed over the project yet. If you want more details about our attempt to move to immutable-js-oss you can join our Slack (https://github.com/immutable-js/immutable-js/issues/1689#issuecomment-672469671). tl;dr We're going to keep maintaining it there until @leebyron gives us access to this repo.