@jdeniau 2021-06-29
Integrated in #1833 via https://github.com/immutable-js/immutable-js/pull/1833/commits/f172650ade3c5d63ce96a45fefa9751b9ae6ff78